Files
Script separado em um arquivo chamado codes.js
